Generalize the inductive exclusion rule, by relaxing the number of cells under consideration, and removing the requirement that their masks be identical.
The more general rule is:

For any N cells which have a total of N possibilities among them, all
N of those numbers must be within those cells, and can't be elsewhere
in the set.

Now the program will search for that condition with N=2, then N=3,
N=4, and so on until it finds something.  This is a relatively
expensive search, and so the program only does this when it runs out
of other tricks.

This allows the solver to solve extreme2.txt and extreme3.txt, where
it previously couldn't.